fix: refactor code and update model semester, course
