+++ README.md
@@ -1,1 +1,17 @@
-123
+# Student Management Server
+
+## Hướng dẫn sử dụng
+
+1. clone dự án từ github về
+
+    ```cmd
+      git clone ...
+    ```
+
+2. tải các package về cho dự án
+
+    ```cmd
+      npm install
+    ```
+
+## Tài liệu api
