+++ src/Model/Course.model.js
@@ -8,12 +8,23 @@ const Course = new Schema({
   code: String,
   className: String,
   time: {
-    jd: Number,
-    shift: String
+    lesson: Number,
+    shift: {
+      type: [Number],
+      enum: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16],
+    },
+    dayOfWeek: {
+      type: [String],
+      enum: ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'],
+    }
   },
-  room: String,
-  tc: Number, 
-  teacher: String
+  semester: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'Semester'
+  }
+}, {
+  timestamps: true,
+  collection: 'courses'
 })
