+++ src/Routes/semesterRouter.js
@@ -2,7 +2,7 @@ import SemesterControler from "@/Controler/semester/SemesterControler";
 import express from "express";
 const router = express.Router()
 const middlewareControler = require('../MiddleWare/middlewareControler')
-const RegisterControler = require('../Controler/register/RegisterControler')
+const RegisterControler = require('../Controller/register/RegisterControler')
 
 router.get('/getAll', SemesterControler.getAll);
 router.get('/get-semester/:userId/:id', SemesterControler.getSemester);
