+++ src/Routes/teacherRouter.js
@@ -1,6 +1,6 @@
 const router = require("express").Router()
 const middlewareControler = require('../MiddleWare/middlewareControler')
-const TeacherControler = require("../Controler/user/TeacherControler")
+const TeacherControler = require("../Controller/user/TeacherControler")
 router.get('/get-all', TeacherControler.getAll)
 router.post('/create-teacher',middlewareControler.verifyTokenIsAdmin, TeacherControler.createTeacher)
